move the excluded terms delegate class
